# contracts/product-api.openapi.yaml
openapi: 3.1.0
info:
  title: PathMemory Product API
  version: 3.0.0
  description: >-
    Public API owned by the Express application backend. Mobile and web call
    this API only. AI perception is internal and never exposed directly to a
    client.
servers:
  - url: https://api.example.invalid
    description: Placeholder only; replace through environment configuration.
...
paths:
  /api/v2/routes:
    post:
      tags: [Routes]
      summary: Create one draft workplace landmark graph
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: Draft route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/routes/{routeId}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [Routes]
      summary: Get a published or draft landmark graph
      operationId: getRoute
      responses:
        '200':
          description: Route returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/routes/{routeId}/sessions:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Sessions]
      summary: Start a learn session or plan a navigation session
      operationId: startRouteSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
            examples:
              navigate:
                externalValue: ./examples/product-navigation-session-request.json
      responses:
        '201':
          description: Session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteSession'
              examples:
                navigationAwaitingStart:
                  externalValue: ./examples/product-navigation-session-success.json
        '409':
          description: Graph/session state is invalid, origin equals destination or no directed path exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/routes/{routeId}/reachable-destinations:
    parameters:
      - $ref: '#/components/parameters/RouteId'
      - name: originLandmarkId
        in: query
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Routes]
      summary: List published landmarks reachable from a selected origin
      operationId: listReachableDestinations
      responses:
        '200':
          description: Screen-reader-friendly destination options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReachableDestinationsResponse'
              examples:
                success:
                  externalValue: ./examples/product-reachable-destinations.json
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Route graph is not published or origin is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/sessions/{sessionId}/observations:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Sessions]
      summary: Analyse sampled frames and return product-safe narration/state
      description: >-
        Express forwards ephemeral image bytes to the internal AI service,
        validates perception, applies deterministic route policy and does not
        persist raw media.
      operationId: createObservation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [clientRequestId, capturedAt, frames]
              properties:
                clientRequestId:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Client-generated ID used to reject stale/duplicate responses.
                capturedAt:
                  type: string
                  format: date-time
                locale:
                  type: string
                  enum: [en-US, vi-VN]
                  default: vi-VN
                  description: Language used for deterministic product narration and AI text extraction.
                frames:
                  type: array
                  minItems: 1
                  maxItems: 3
                  description: Per-observation frame batch limit; unrelated to total route landmarks.
                  items:
                    type: string
                    format: binary
                sensorEvidence:
                  $ref: '#/components/schemas/SensorEvidence'
            encoding:
              frames:
                contentType: image/jpeg, image/png
              sensorEvidence:
                contentType: application/json
      responses:
        '200':
          description: Product-safe observation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationProductResponse'
              examples:
                learnCandidate:
                  externalValue: ./examples/product-observation-success.json
                navigationStartConfirmed:
                  externalValue: ./examples/product-navigation-start-confirmed.json
                navigationStartNotConfirmed:
                  externalValue: ./examples/product-navigation-start-not-confirmed.json
        '413':
          description: Frame count or payload exceeds configured limits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/AiUnavailable'
        '504':
          $ref: '#/components/responses/AiTimeout'
  /api/v2/sessions/{sessionId}/landmarks:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Sessions]
      summary: Save a useful unique candidate as an AI draft landmark
      operationId: saveLandmarkDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLandmarkDraftRequest'
      responses:
        '201':
          description: Draft landmark saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Landmark'
        '409':
          description: Candidate is a deterministic duplicate of an existing route landmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/sessions/{sessionId}/finish:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Sessions]
      summary: Finish a learn or navigate session
      operationId: finishSession
      responses:
        '200':
          description: Session finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteSession'
        '409':
          description: Session cannot be finished in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/routes/{routeId}/landmarks/{landmarkId}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
      - $ref: '#/components/parameters/LandmarkId'
    patch:
      tags: [Review]
      summary: Edit and verify a landmark draft as a human admin/buddy
      operationId: reviewLandmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewLandmarkRequest'
      responses:
        '200':
          description: Landmark updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Landmark'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/routes/{routeId}/edges:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    put:
      tags: [Review]
      summary: Replace human-reviewed directed relations between landmarks
      description: >-
        Used by the admin/buddy review UI. Direction belongs to the directed
        edge between two landmarks, not to either landmark itself. Express
        rejects self-loops, duplicate directed edges and references to
        landmarks outside the graph. Branching and reverse edges are allowed.
      operationId: replaceRouteEdges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceRouteEdgesRequest'
            examples:
              boundedGraphDemo:
                externalValue: ./examples/product-route-edges-request.json
      responses:
        '200':
          description: Route directions replaced and validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '409':
          description: Edge set contains invalid references, self-loops or duplicate directed edges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/routes/{routeId}/publish:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Review]
      summary: Publish a fully human-verified landmark graph
      operationId: publishRoute
      responses:
        '200':
          description: Route published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '409':
          description: Graph has fewer than two landmarks, unverified landmarks or invalid topology
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/routes/{routeId}/mark-outdated:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Review]
      summary: Prevent further replay until a human revalidates the route
      operationId: markRouteOutdated
      responses:
        '200':
          description: Route marked outdated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
components:
  parameters:
    RouteId:
      name: routeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              externalValue: ./examples/product-validation-error.json
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    CreateRouteRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        locale:
          type: string
          default: vi-VN
          maxLength: 20
    StartSessionRequest:
      oneOf:
        - $ref: '#/components/schemas/StartLearnSessionRequest'
        - $ref: '#/components/schemas/StartNavigateSessionRequest'
      discriminator:
        propertyName: mode
...
    SensorEvidence:
      type: object
      additionalProperties: false
      properties:
        headingDeltaDegrees:
          type: number
          minimum: -360
          maximum: 360
          description: Supporting relative rotation evidence; never treated as absolute indoor position.
        stepDelta:
          type: integer
          minimum: 0
          maximum: 1000
          description: Supporting progress evidence; never presented as exact distance.
    ObservationProductResponse:
      type: object
      additionalProperties: false
      required:
        - requestId
        - observationId
        - sessionMode
        - routeState
        - landmarkMatchStatus
        - spokenMessage
        - shouldAdvance
      properties:
        requestId:
          type: string
        observationId:
          type: string
          format: uuid
        sessionMode:
          type: string
          enum: [LEARN, NAVIGATE]
        routeState:
          type: string
          enum: [OBSERVING, AWAITING_START_CONFIRMATION, SEEKING_LANDMARK, STOP_AND_RESCAN, ROUTE_COMPLETED]
        landmarkMatchStatus:
          type: string
          enum: [NOT_APPLICABLE, CANDIDATE, MATCHED, NOT_MATCHED, INSUFFICIENT_EVIDENCE]
        candidateLandmark:
          oneOf:
            - $ref: '#/components/schemas/CandidateLandmark'
            - type: 'null'
        expectedLandmark:
          oneOf:
            - $ref: '#/components/schemas/LandmarkSummary'
            - type: 'null'
        spokenMessage:
          type: string
          minLength: 1
          maxLength: 500
        shouldAdvance:
          type: boolean
          description: Deterministic Express decision; never copied from model output.
        retryAllowed:
          type: boolean
          default: true
...
    Route:
      type: object
      additionalProperties: false
      description: >-
        A bounded workplace landmark graph. A concrete origin-to-destination
        path is computed when a NAVIGATE session starts.
      required: [id, name, status, landmarks, edges, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OUTDATED]
        landmarks:
          type: array
          minItems: 0
          items:
            $ref: '#/components/schemas/Landmark'
        edges:
          type: array
          minItems: 0
          description: >-
            Directed human-reviewed relative instructions. Draft graphs may
            omit edges; published graphs may branch and include reverse edges.
          items:
            $ref: '#/components/schemas/RouteEdge'
        createdAt:
          type: string
          format: date-time
        publishedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    RouteSession:
      type: object
      additionalProperties: false
      required: [id, routeId, mode, status, startedAt]
      properties:
        id:
          type: string
          format: uuid
        routeId:
          type: string
          format: uuid
        mode:
          type: string
          enum: [LEARN, NAVIGATE]
        status:
          type: string
          enum: [ACTIVE, AWAITING_START_CONFIRMATION, COMPLETED, CANCELLED]
        currentPathIndex:
          type: integer
          minimum: 0
          description: Index in plannedPath.landmarkIds; zero is the selected origin awaiting confirmation.
        originLandmarkId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        destinationLandmarkId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        expectedLandmarkId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        plannedPath:
          oneOf:
            - $ref: '#/components/schemas/NavigationPath'
            - type: 'null'
        startedAt:
          type: string
          format: date-time
        finishedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - INVALID_STATE
                - DUPLICATE_LANDMARK
                - INVALID_ROUTE_TOPOLOGY
                - NO_ROUTE_AVAILABLE
                - AI_INVALID_RESPONSE
                - AI_PROVIDER_UNAVAILABLE
                - AI_TIMEOUT
                - PAYLOAD_TOO_LARGE
            message:
              type: string
            retryable:
              type: boolean
            requestId:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, issue]
                properties:
                  field:
                    type: string
                  issue:
                    type: string
...
    ReachableDestinationsResponse:
      type: object
      additionalProperties: false
      required: [routeId, origin, destinations]
      properties:
        routeId:
          type: string
          format: uuid
        origin:
          $ref: '#/components/schemas/LandmarkSummary'
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/LandmarkSummary'
...